# Conflicts:
#	app/src/main/java/com/team/app/ui/home/HomePageViewModel.kt

+++ app/src/main/java/com/team/app/ui/home/HomePageViewModel.kt
@@ -1,6 +1,7 @@
 package com.team.app.ui.home
 
 import androidx.compose.runtime.mutableIntStateOf
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import com.lonartie.bookdiary.data.repositories.SettingsRepository
@@ -13,9 +14,9 @@ import com.team.app.data.repositories.StepCounterRepository
 import com.team.app.data.repositories.AttributesRepository
 import com.team.app.data.repositories.HotbarRepository
 import com.team.app.data.repositories.InventoryRepository
+import com.team.app.utils.Constants.Companion.INVALID_INVENTORY_ITEM
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
 import javax.inject.Inject
 import kotlin.math.min
 
@@ -31,10 +32,14 @@ class HomePageViewModel @Inject constructor(
     val figureState = mutableIntStateOf(R.drawable.figure_happy)
     val attributes: Flow<Attributes> = attributesRepo.getAttributes()
 
-    val hotbar: Flow<Hotbar> = inventoryRepo.getItemsFlow()
-        .map {
-            hotbarRepo.getHotbar()
-        }
+    val hotbar = mutableStateOf(
+        Hotbar(
+            INVALID_INVENTORY_ITEM,
+            INVALID_INVENTORY_ITEM,
+            INVALID_INVENTORY_ITEM
+        )
+    )
+    val inventoryItems = inventoryRepo.getItemsFlow()
 
     private fun getMinimalAttributeValue(attributes: Attributes): Int {
         return min(
@@ -47,45 +52,29 @@ class HomePageViewModel @Inject constructor(
 
     }
 
+    suspend fun updateHotbar() {
+        hotbar.value = hotbarRepo.getHotbar()
+    }
+
     fun setFigureState(attributes: Attributes) {
-        val minimalValue =
-            getMinimalAttributeValue(attributes)
-        figureState.intValue =
-            when {
-                minimalValue >= 75 -> R.drawable.figure_happy
-                minimalValue >= 50 -> R.drawable.figure_lonely
-                else -> R.drawable.figure_angry
-            }
+        val minimalValue = getMinimalAttributeValue(attributes)
+        figureState.intValue = when {
+            minimalValue >= 75 -> R.drawable.figure_happy
+            minimalValue >= 50 -> R.drawable.figure_lonely
+            else -> R.drawable.figure_angry
+        }
     }
 
     suspend fun onStart() {
-        var item =
-            Item(
-                ItemType.FOOD,
-                "Chicken",
-                10,
-                10
-            )
+        var item = Item(ItemType.FOOD, "Chicken", 10, 10)
         println("current food: $item")
         settings.saveCurrentItem(item)
 
-        item =
-            Item(
-                ItemType.TOY,
-                "Ball",
-                5,
-                5
-            )
+        item = Item(ItemType.TOY, "Ball", 5, 5)
         println("current toy: $item")
         settings.saveCurrentItem(item)
 
-        item =
-            Item(
-                ItemType.MEDICINE,
-                "Medicine",
-                15,
-                15
-            )
+        item = Item(ItemType.MEDICINE, "Medicine", 15, 15)
         println("current misc: $item")
         settings.saveCurrentItem(item)
 
@@ -115,9 +104,12 @@ class HomePageViewModel @Inject constructor(
     }
 
     suspend fun giveFood(item: Item, attributes: Attributes) {
+        if (item.name == "") return
+
         attributesRepo.updateHunger(attributes.hunger + item.actionValue)
         inventoryRepo.removeOne(item)
         setFigureState(attributes)
+        updateHotbar()
     }
 
     suspend fun selectFood() {
@@ -125,9 +117,12 @@ class HomePageViewModel @Inject constructor(
     }
 
     suspend fun giveToy(item: Item, attributes: Attributes) {
+        if (item.name == "") return
+
         attributesRepo.updateHappiness(attributes.happiness + item.actionValue)
         inventoryRepo.removeOne(item)
         setFigureState(attributes)
+        updateHotbar()
     }
 
     suspend fun selectToy() {
@@ -135,10 +130,13 @@ class HomePageViewModel @Inject constructor(
     }
 
     suspend fun giveItem(item: Item, attributes: Attributes) {
+        if (item.name == "") return
+
         if (item.itemType == ItemType.MEDICINE) {
             attributesRepo.updateHealth(attributes.health + item.actionValue)
             inventoryRepo.removeOne(item)
             setFigureState(attributes)
+            updateHotbar()
         }
     }
